+++ src/cl.accordion/cl.accordion.js
@@ -1,7 +1,7 @@
 /*!
  * @author      Angelo Dini - github.com/finalangel/classjs-plugins
  * @copyright	Distributed under the BSD License.
- * @version     1.1.0
+ * @version     1.1.1
  */
 
 // ensure namespace is defined
@@ -18,18 +18,24 @@ var Cl = window.Cl || {};
 		 */
 
 		options: {
-			'index': null, // first element to show
-			'expanded': false, // show element on initialization
-			'event': 'click', // jquery event
+			'index': null,
+			'expanded': false,
+			'event': 'click focusin',
 			'easing': 'linear',
-			'duration': 300, // animation duration
-			'grouping': true, // enables only one item to be opened at a time
-			'forceClose': false, // if true closes current element when clicking again
+			'duration': 300,
+			'grouping': true,
+			'forceClose': false,
+			'disableAnchors': true,
 			'cls': {
-				'expanded': 'expanded', // class that will be added
-				'collapsed': 'collapsed', // class that will be added
+				'expanded': 'expanded',
+				'collapsed': 'collapsed',
 				'trigger': '.trigger',
-				'container': '.container'
+				'container': '.container',
+				'text': '.text'
+			},
+			'lang': {
+				'expanded': 'Expanded ',
+				'collapsed': 'Collapsed '
 			}
 		},
 
@@ -51,106 +57,114 @@ var Cl = window.Cl || {};
 		_setup: function () {
 			var that = this;
 
-			// set initial trigger state
-			this.triggers.addClass(this.options.cls.collapsed);
-
-			// we need to set the heights for each element
+			// calculate height for each item
 			this.containers.each(function (index, item) {
-				$(item).height($(item).outerHeight(true));
+				$(item).height($(item).height());
 			});
 
-			this.triggers.bind(this.options.event, function (e) {
+			// add event to each trigger
+			this.triggers.on(this.options.event, function (e) {
 				e.preventDefault();
 				that.toggle(that.triggers.index(this));
 			});
 
-			// check if containers are expanded
+			// prevent click events on substitutes
+			if(this.options.disableAnchors) {
+				this.triggers.find('a', function (e) {
+					e.preventDefault();
+				});
+			}
+
+			// setup initial states
 			if(this.options.expanded) {
 				this.containers.show();
 				this.triggers.addClass(this.options.cls.expanded);
+				this.triggers.find('.text').html(this.options.lang.expanded);
+				this.containers.attr('aria-expanded', true);
 			} else {
 				this.containers.hide();
+				this.triggers.addClass(this.options.cls.collapsed);
+				this.triggers.find('.text').html(this.options.lang.collapsed);
+				this.containers.attr('aria-expanded', false);
 			}
 
-			// show correct index
+			// if index is defined and the elements are not expanded, show provided index
 			if(this.options.index !== null && !this.options.expanded) this.toggle(this.options.index);
 		},
 
 		toggle: function (index) {
 			// trigger event
-			this._triggerEvent('toggle');
+			this._fire('toggle');
 
-			// cancel if index is the same
+			// cancel if index is the same end forceClose disabled
 			if(this.index === index && !this.options.forceClose) return false;
 
 			// set global index
 			this.index = index;
 
-			// hide containers if grouping is active
-			if(this.options.grouping) this.hide();
-
-			// remove all classes
+			// if grouping is enabled
 			if(this.options.grouping) {
+				// hide all containers
+				this.hide();
+				// reset triggers to initial states
 				this.triggers.removeClass(this.options.cls.expanded)
-					.addClass(this.options.cls.collapsed);
+					.addClass(this.options.cls.collapsed)
+					.find('.text').html(this.options.lang.collapsed);
 			}
 
 			// redirect to required behaviour
 			(this.containers.eq(index).is(':visible')) ? this.hide(index) : this.show(index);
 
 			// trigger callback
-			this._triggerCallback('toggle', this);
+			this._fire('toggle', this);
 		},
 
 		show: function (index) {
 			// trigger event
-			this._triggerEvent('show');
-
-			var container = this.containers.eq(index);
-			var trigger = this.triggers.eq(index);
+			this._fire('show');
 
 			// if no index is provided, hide all
-			if(index === undefined) this.containers.slideDown(this.options.duration, this.options.easing);
-
-			// do animation
-			container.slideDown(this.options.duration, this.options.easing);
-
-			// add and remove classes
-			trigger.addClass(this.options.cls.expanded).removeClass(this.options.cls.collapsed);
+			if(index === undefined) {
+				this.containers.slideDown(this.options.duration, this.options.easing);
+			// otherwise show element
+			} else {
+				this.containers.eq(index).slideDown(this.options.duration, this.options.easing);
+				this.triggers.eq(index).addClass(this.options.cls.expanded).removeClass(this.options.cls.collapsed)
+					.find('.text').html(this.options.lang.expanded);
+			}
 
 			// trigger callback
-			this._triggerCallback('show', this);
+			this._fire('show', this);
 		},
 
 		hide: function (index) {
 			// trigger event
-			this._triggerEvent('hide');
-
-			var container = this.containers.eq(index);
-			var trigger = this.triggers.eq(index);
+			this._fire('hide');
 
 			// if no index is provided, hide all
-			if(index === undefined) this.containers.slideUp(this.options.duration, this.options.easing);
-
-			// do animation
-			container.slideUp(this.options.duration, this.options.easing);
-
-			// add and remove classes
-			trigger.addClass(this.options.cls.collapsed).removeClass(this.options.cls.expanded);
+			if(index === undefined) {
+				this.containers.slideUp(this.options.duration, this.options.easing);
+			// otherwise hide element
+			} else {
+				this.containers.eq(index).slideUp(this.options.duration, this.options.easing);
+				this.triggers.eq(index).addClass(this.options.cls.collapsed).removeClass(this.options.cls.expanded)
+					.find('.text').html(this.options.lang.collapsed);
+			}
 
 			// trigger callback
-			this._triggerCallback('hide', this);
+			this._fire('hide', this);
 		},
 
-		_triggerCallback: function (fn, scope) {
-			// cancel if there is no callback found
-			if(this.callbacks[fn] === undefined) return false;
-			// excecute fallback
-			this.callbacks[fn](scope);
-		},
-
-		_triggerEvent: function (event) {
-			$.event.trigger('accordion-' + event);
+		_fire: function (keyword, scope) {
+			if(scope) {
+				// cancel if there is no callback found
+				if(this.callbacks[keyword] === undefined) return false;
+				// excecute callback
+				this.callbacks[fn](scope);
+			} else {
+				// excecute event
+				$.event.trigger('accordion-' + event);
+			}
 		}
 
 	});